refactor(modules): move app resource from src to module
