+++ src/modules/app/index.ts
@@ -1,0 +1,1 @@
+export { AppModule } from './app.module';
